Adding 'new logo' and generalizing more components
