--- src/components/NormalLabel.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components/native'
-
-export default NormalLabel = styled.Text`
-    width: 80%;
-    font-family: Trueno-SemiBold;
-    font-size: 16px;
-    color: #011640;
-    text-align: left;
-    margin-left: 20px;
-    margin-bottom: 15px;
-`;
