--- src/components/RedStatusBar.js
@@ -1,9 +1,0 @@
-import { StatusBar } from 'expo-status-bar'
-
-import styled from 'styled-components/native'
-
-export default RedBar = styled(StatusBar).attrs({
-    barStyle: "light-content",
-    backgroundColor: "#E83338",
-    translucent: false
-})``
